Moved HarryRunner interface to harry-core and removed the harry-runner module.
