+++ harry-core/src/harry/runner/HarryRunner.java
@@ -18,6 +18,13 @@
 
 package harry.runner;
 
+import harry.core.Configuration;
+import harry.core.Run;
+import harry.corruptor.*;
+import harry.util.ThrowingRunnable;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.Random;
@@ -25,19 +32,6 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import harry.core.Configuration;
-import harry.core.Run;
-import harry.corruptor.AddExtraRowCorruptor;
-import harry.corruptor.ChangeValueCorruptor;
-import harry.corruptor.HideRowCorruptor;
-import harry.corruptor.HideValueCorruptor;
-import harry.corruptor.QueryResponseCorruptor;
-import harry.model.ExhaustiveChecker;
-import harry.model.OpSelectors;
 
 public interface HarryRunner
 {
