+++ pom.xml
@@ -33,7 +33,6 @@
         <module>harry-core</module>
         <module>harry-integration</module>
         <module>harry-integration-external</module>
-        <module>harry-runner</module>
     </modules>
 
     <properties>
